+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG.md
 
-## [0.1.0] - 19.11.2025
-- Added naming boards
+## [0.3.0] - 25.11.2025
+- Added editing board name
 
 ## [0.2.0] - 19.11.2025
 - Added show all boards and show one board
 - Changed from one single connection and cursor of BoardManager to separate connections and cursors for each operation
 - Added singal handler in run_bot.py
+
+## [0.1.0] - 19.11.2025
+- Added naming boards
+